New routing system for admin based on roles
