+++ MakersMarkt/app/Http/Controllers/Admin/ProductController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\Product;
 
 class ProductController extends Controller
 {
-    public function show($id) {
-        //
+    public function index() {
+        $products = Product::all();
+        return view('admin.products', compact('products'));
     }
 }
