+++ MakersMarkt/app/Http/Controllers/Admin/UserController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+use App\Models\User;
+
+class UserController extends Controller
+{
+    public function index() {
+        $users = User::where('verified', false)->get();
+        return view('admin.users', compact('users'));
+    }
+}
